refactor: Extract signing key tests.
