+++ integration/simple-cache-client.test.ts
@@ -2,10 +2,7 @@ import {v4} from 'uuid';
 import {
   CollectionTtl,
   CacheDelete,
-  CreateSigningKey,
-  ListSigningKeys,
   MomentoErrorCode,
-  RevokeSigningKey,
   CacheDictionaryFetch,
   CacheDictionarySetField,
   CacheDictionarySetFields,
@@ -19,47 +16,6 @@ import {TextEncoder} from 'util';
 import {SetupIntegrationTest} from './integration-setup';
 
 const {Momento, IntegrationTestCacheName} = SetupIntegrationTest();
-
-describe('Signing keys', () => {
-  it('should create, list, and revoke a signing key', async () => {
-    const createSigningKeyResponse = await Momento.createSigningKey(30);
-    expect(createSigningKeyResponse).toBeInstanceOf(CreateSigningKey.Success);
-    let listSigningKeysResponse = await Momento.listSigningKeys();
-    expect(listSigningKeysResponse).toBeInstanceOf(ListSigningKeys.Success);
-    let signingKeys = (
-      listSigningKeysResponse as ListSigningKeys.Success
-    ).getSigningKeys();
-    expect(signingKeys.length).toBeGreaterThan(0);
-    expect(
-      signingKeys
-        .map(k => k.getKeyId())
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-        .some(
-          k =>
-            k ===
-            (createSigningKeyResponse as CreateSigningKey.Success).getKeyId()
-        )
-    ).toEqual(true);
-    const revokeResponse = await Momento.revokeSigningKey(
-      (createSigningKeyResponse as CreateSigningKey.Success).getKeyId()
-    );
-    expect(revokeResponse).toBeInstanceOf(RevokeSigningKey.Success);
-    listSigningKeysResponse = await Momento.listSigningKeys();
-    expect(listSigningKeysResponse).toBeInstanceOf(ListSigningKeys.Success);
-    signingKeys = (
-      listSigningKeysResponse as ListSigningKeys.Success
-    ).getSigningKeys();
-    expect(
-      signingKeys
-        .map(k => k.getKeyId())
-        .some(
-          k =>
-            k ===
-            (createSigningKeyResponse as CreateSigningKey.Success).getKeyId()
-        )
-    ).toEqual(false);
-  });
-});
 
 describe('Integration tests for dictionary operations', () => {
   it('should return InvalidArgument response for dictionaryGetField with invalid cache and dictionary names', async () => {
